Use github token
